pre-sliding window preparations
this solved a bug, that I was actually pretty dumb in how I found the
best matching unit, so this was perhaps fixed,
I also implemented unshuffling and changed the way that the shuffling
was done. it adds an extra step to each iteration, but it also enables
for intermediate layers to be shuffled.